refactor: LoginChecjPage추가 및 navigate수정
