+++ src/router/Routes.tsx
@@ -3,6 +3,7 @@ import { ROUTE_PATH } from '@/constants/routes';
 import HomePage from '@/pages/Home/HomePage';
 import LoginPage from '@/pages/Login/LoginPage';
 import LoginCallback from '@/pages/Login/components/LoginCallback';
+import LoginCheckPage from '@/pages/Login/LoginCheckPage';
 import MyPage from '@/pages/My/MyPage';
 import PickPage from '@/pages/Pick/PickPage';
 import FestivalsPage from '@/pages/Festivals/FestivalsPage';
@@ -13,6 +14,7 @@ const Routes = () => {
       <Route path={ROUTE_PATH.HOME} element={<HomePage />} />
       <Route path={ROUTE_PATH.LOGIN} element={<LoginPage />} />
       <Route path={ROUTE_PATH.LOGIN_CALLBACK} element={<LoginCallback />} />
+      <Route path="/login/check" element={<LoginCheckPage />} />
       <Route path={ROUTE_PATH.MY} element={<MyPage />} />
       <Route path={ROUTE_PATH.PICK} element={<PickPage />} />
       <Route path={ROUTE_PATH.FESTIVALS} element={<FestivalsPage />} />
